+++ CHANGELOG.md
@@ -1,6 +1,13 @@
 # Changelog
 <!-- https://keepachangelog.com/en/1.0.0/ -->
 
+## [0.0.2]  2022-07-17
+### Added
+- documentation
+- make Label a class
+### Breaking Change
+- match step takes spread array instead of array
+
 ## [0.0.1]  2022-07-17
 ### Added
 - initial project